+++ Core/SyncDataProviders.swift
@@ -100,17 +100,17 @@ public class SyncDataProviders: DataProvidersSource {
         secureVaultErrorReporter: SecureVaultReporting,
         settingHandlers: [SettingSyncHandler],
         favoritesDisplayModeStorage: FavoritesDisplayModeStoring,
-                    syncErrorHandler: SyncErrorHandling
+        syncErrorHandler: SyncErrorHandling
     ) {
         self.bookmarksDatabase = bookmarksDatabase
         self.secureVaultFactory = secureVaultFactory
         self.secureVaultErrorReporter = secureVaultErrorReporter
         bookmarksAdapter = SyncBookmarksAdapter(database: bookmarksDatabase,
                                                 favoritesDisplayModeStorage: favoritesDisplayModeStorage,
-                                                syncErrorHandler:             syncErrorHandler)
+                                                syncErrorHandler: syncErrorHandler)
         credentialsAdapter = SyncCredentialsAdapter(secureVaultFactory: secureVaultFactory,
                                                     secureVaultErrorReporter: secureVaultErrorReporter,
-                                                    syncErrorHandler:             syncErrorHandler)
+                                                    syncErrorHandler: syncErrorHandler)
         settingsAdapter = SyncSettingsAdapter(settingHandlers: settingHandlers)
     }
