+++ DuckDuckGoTests/CapturingSyncPausedStateManager.swift
